# doc/api.yaml
openapi: 3.0.3
info:
  title: Simplified login API specification
  description: |-
    This OpenAPI document describes the simplified login API.
    Copy and paste the API from the `paths` key to your OpenAPI document.
  version: "1"
...
paths:
...
  /users/{userId}/logo:
    get:
      parameters:
        - name: userId
          schema:
            type: integer
            description: User identifier

          in: path
          required: true
          description: userId

      tags: ["Photos"]
      summary: Get logo of a user specified by parameter
      operationId: getLogo
      description: |
        Obtains the raw image. If it does not exist, it returns -1.

      responses:
        "200":
          description: Logo identifier
          content:
            text/plain:
              schema:
                type: integer
                example: 9013213
                description: Logo's identifier


        '400':
          description: Access token is missing or invalid

        '404':
          description: 404 Not found
...
  /users/{userId}/commentPhoto/{photoId}:
    post:
      parameters:
        - name: userId
          schema:
            type: integer
            description: User identifier

          in: path
          required: true
          description: userId

        - name: photoId
          schema:
            type: integer
            description: Photo identifier

          in: path
          required: true
          description: ID of the photo to be commented




      tags: ["Photos"]
      summary: Comments a photo of a user
      description: |-
        The logged user comments a photo of another user. Each photo has a unique identifier, so we won't have to specify the user.

      operationId: commentPhoto
      requestBody:
        description: Comment to be added to the requested photo.
        content:

          text/plain:
            schema:
              type: string
              description: Photo comment.
              example: "Wow such a nice picture"
              pattern: ^[A-Z][0-9]$
              minLength: 3
              maxLength: 144


        required: true


      responses:
        "201":
          description: Posted comment
          content:
            application/json:
              schema:
                description: Posted comment schema
                type: object
                properties:

                  id:
                    type: integer
                    description: ID of the comment
                    example: 9123

                  content:
                    type: string
                    description: Photo comment.
                    example: "Wow, such a nice picture!"
                    minLength: 3
                    maxLength: 144

                  photoId:
                    type: integer
                    description: ID of the photo commented
                    example: 131

                  userId:
                    type: integer
                    description: ID of the user that commented
                    example: 1

                  date:
                    type: string
                    description: Date in which the photo was uploaded
                    minLength: 3
                    maxLength: 16
                    format: date-time




        '400':
          description: Bad Request. Error parsing data or information not valid.

        '401':
          description: Unauthorized. Missing or invalid Authentication token.



        '405':
          description: Method not allowed.
          content:
            text/plain:
              schema:
                type: string
                example: "Method Not Allowed"
                description: Error description
  /users/{userId}/likePhoto/{photoId}:
    put:
      parameters:
        - name: userId
          schema:
            type: integer
            description: User identifier

          in: path
          required: true
          description: userId

        - name: photoId
          schema:
            type: integer
            description: Photo identifier

          in: path
          required: true
          description: ID of the photo to be liked




      tags: ["Photos"]
      summary: Likes a photo of a user
      description: |-
        The logged user likes a photo of another user. Each photo has a unique identifier, so we won't have to specify the user.

      operationId: likePhoto
      

      responses:
        '200':
          description: Updated image
          content:
            application/json:
              schema:
                description: Updated image schema
                type: object
                properties:

                  imagePath:
                    description: path to the image
                    type: string
                    minLength: 3
                    maxLength: 100

                  photoID:
                    type: integer
                    description: ID of the photo
                    example: 9123

                  likes:
                    type: array
                    description: An array with the likes of each photo
                    minItems: 0
                    maxItems: 100
                    items:
                      type: integer
                      description: It stores the IDs of the users that liked the photo
                      example: 7



                  comments:
                    type: array
                    description: An array with the comments of each photo
                    minItems: 0
                    maxItems: 100


                    items:
                      type: object
                      properties:
                        user:
                          type: integer
                          description: User ID
                          example: 7


                        comment:
                          type: string
                          description: Comment
                          minLength: 3
                          maxLength: 16
                          pattern: ^[A-Z][0-9]$


                  date:
                    type: string
                    description: Date in which the photo was uploaded
                    minLength: 3
                    maxLength: 16
                    format: date-time
        '400':
          description: Bad Request. Error parsing data or information not valid.


        '401':
          description: Unauthorized. Missing or invalid Authentication token.


        '405':
          description: Method not allowed.
          content:
            text/plain:
              schema:
                type: string
                example: "Method Not Allowed"
                description: Error description
...
  /users/{userId}/followUser/{userFollowed}:
    put:
      parameters:
        - name: userId
          schema:
            type: integer
            description: User identifier

          in: path
          required: true
          description: userId

        - name: userFollowed
          schema:
            type: integer
            description: User identifier

          in: path
          required: true
          description: user to be followed


      tags: ["User"]
      summary: Given a user ID, the logged user follows it.
      description: |-
        The logged user follows the user identified by the parameter. It returns 0 if it has been correctly followed, or -1 if an error has occurred (already followed or not found).

      operationId: followUser


      responses:
        '200':
          description: Information related to the user
          content:
            application/json:
              schema:
                description: Information of the user; username, profile pic, followers,photos
                type: object
                properties:

                  id:
                    description: User identifier
                    type: integer
                    example: 9013213

                  name:
                    type: string
                    example: "Maria"
                    minLength: 3
                    maxLength: 16
                    pattern: ^[a-zA-Z0-9]{1,20}$
                    description: Username

                  profilePictureID:
                    description: Profile picture identifier
                    type: integer
                    example: 123912

                  banned:
                    type: array
                    description: An array with the bannedUsers.
                    minItems: 0
                    maxItems: 100
                    items:
                      type: string
                      minLength: 3
                      maxLength: 16

                  followers:
                    type: array
                    description: An array with the followers. We only store the followers, not the followed.
                    minItems: 0
                    maxItems: 100
                    items:
                      type: integer
                      description: ID of the user that follows



                  photos:
                    type: array
                    description: An array with the stream of photos.
                    minItems: 0
                    maxItems: 100
                    items:
                      type: object
                      properties:

                        imagePath:
                          description: path to the image
                          type: string
                          minLength: 3
                          maxLength: 100


                        photoID:
                          description: ID of the photo
                          type: integer

                        likes:
                          type: array
                          description: A sub-array with the likes of each photo.
                          minItems: 0
                          maxItems: 100
                          items:
                            type: integer
                            description: ID of the user that liked the photo


                        comments:
                          type: array
                          description: A sub-array with the comments of each photo.
                          minItems: 0
                          maxItems: 100
                          items:
                            type: object
                            properties:

                              user:
                                type: integer
                                description: User ID


                              comment:
                                type: string
                                minLength: 3
                                maxLength: 16
                                description: comment


                        date:
                          type: string
                          minLength: 3
                          maxLength: 16
                          format: date-time
                          description: Creation date and time


        '400':
          description: Bad Request. Error parsing data or information not valid.


        '405':
          description: Method not allowed.
          content:
            text/plain:
              schema:
                type: string
                example: "Method Not Allowed"
                description: Error description
...
  /users/{userId}/deletePhoto/{photoId}:
    delete:
      parameters:
        - name: userId
          schema:
            type: integer
            description: User identifier
          in: path
          required: true
          description: user logged

        - name: photoId
          schema:
            type: integer
            description: Photo identifier

          in: path
          required: true
          description: ID of the photo to be deleted




      tags: ["Photos"]
      summary: Given a user ID and a photo Id, the logged user deletes it from its stream
      description: |-
        The logged user deletes the photo from its stream. 
        It returns 0 if it has been correctly deleted, or -1 if an error has occurred (userId or photoId not found).

      operationId: deletePhoto


      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: integer
                example: 1
                description: Successful operation


        '400':
          description: Access token is missing or invalid

        '404':
          description: Resource not found.
          content:
            application/json:
              schema:
                type: integer
                example: -1
                description: Non successful operation
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
security:
  - bearerAuth: []